+++ noodles-gff/src/record.rs
@@ -363,6 +363,8 @@ mod tests {
 
     #[test]
     fn test_from_str() -> Result<(), Box<dyn std::error::Error>> {
+        use self::attributes::field::{Key, Value};
+
         let s = "sq0\tNOODLES\tgene\t8\t13\t.\t+\t.\tgene_id=ndls0;gene_name=gene0";
         let record = s.parse::<Record>()?;
 
@@ -377,10 +379,12 @@ mod tests {
 
         assert_eq!(
             record.attributes(),
-            &Attributes::from(vec![
-                attributes::Entry::new("gene_id", "ndls0"),
-                attributes::Entry::new("gene_name", "gene0"),
-            ])
+            &[
+                (Key::from("gene_id"), Value::from("ndls0")),
+                (Key::from("gene_name"), Value::from("gene0")),
+            ]
+            .into_iter()
+            .collect()
         );
 
         Ok(())